hw2 base part ref

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -6,10 +6,7 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
@@ -18,11 +15,12 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 @Repository
 public class InMemoryMealRepository implements MealRepository {
+    private static final Logger log = getLogger(InMemoryMealRepository.class);
     private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private final AtomicInteger counter = new AtomicInteger(0);
-    private static final Logger log = getLogger(InMemoryMealRepository.class);
+
     {
-        MealsUtil.meals.forEach(x-> save(x, x.getUserId()));
+        MealsUtil.meals.forEach(x -> save(x, x.getUserId()));
     }
 
     @Override
@@ -41,25 +39,29 @@ public class InMemoryMealRepository implements MealRepository {
     @Override
     public boolean delete(int id, int userId) {
         log.info("delete {}", id);
-        if (repository.get(id).getUserId()== userId)
+        if(!repository.containsKey(id)) return false;
+        if (repository.get(id).getUserId() == userId) {
             return repository.remove(id) != null;
-        else
+        }else {
             return false;
+        }
     }
 
     @Override
     public Meal get(int id, int userId) {
         log.info("get {}", id);
-        if (repository.get(id).getUserId()== userId)
+        if (!repository.containsKey(id)) return null;
+        if (repository.get(id).getUserId() == userId) {
             return repository.get(id);
-        else
+        }else {
             return null;
+        }
     }
 
     @Override
-    public Collection<Meal> getAll(int userId) {
+    public List<Meal> getAll(int userId) {
         log.info("getAll");
-        return repository.values().isEmpty() ? Collections.emptyList() : repository.values().stream().filter(x->x.getUserId()==userId)
+        return repository.values().stream().filter(x -> x.getUserId() == userId)
                 .sorted(Comparator.comparing(Meal::getDateTime).reversed()).collect(Collectors.toList());
     }
 }
